+++ intelmq/bots/experts/idea/expert.py
@@ -41,8 +41,6 @@ class IdeaExpertBot(Bot):
         "defacement": "Intrusion.AppCompromise",
         "compromised": "Intrusion.AdminCompromise",
         "backdoor": "Intrusion.AdminCompromise",
-        "vulnerable-service": "Vulnerable.Open",
-        "vulnerable-client": "Vulnerable.Config",
         "blacklist": "Other",
         "dga domain": "Anomaly.Behaviour",
         "proxy": "Vulnerable.Config",