+++ intelmq/bots/parsers/shadowserver/config.py
@@ -221,7 +221,7 @@ open_db2_discovery_service = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-db2-discovery-service',
     }
 }
@@ -296,7 +296,7 @@ open_mdns = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-mdns',
         'protocol.application': 'mdns',
     }
@@ -324,7 +324,7 @@ open_chargen = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-chargen',
         'protocol.application': 'chargen',
     },
@@ -355,7 +355,7 @@ open_tftp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-tftp',
         'protocol.application': 'tftp',
     },
@@ -526,7 +526,7 @@ open_redis = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-redis',
         'protocol.application': 'redis',
     },
@@ -556,7 +556,7 @@ open_portmapper = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-portmapper',
         'protocol.application': 'portmapper',
     },
@@ -600,7 +600,7 @@ open_ipmi = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-ipmi',
         'protocol.application': 'ipmi',
         'protocol.transport': 'udp',
@@ -629,7 +629,7 @@ open_qotd = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-qotd',
         'protocol.application': 'qotd',
     },
@@ -666,7 +666,7 @@ open_ssdp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-ssdp',
         'protocol.application': 'ssdp',
     },
@@ -695,7 +695,7 @@ open_snmp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-snmp',
         'protocol.application': 'snmp',
     },
@@ -729,7 +729,7 @@ open_mssql = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-mssql',
         'protocol.application': 'mssql',
     },
@@ -766,7 +766,7 @@ open_mongodb = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-mongodb',
         'protocol.application': 'mongodb',
     },
@@ -797,7 +797,7 @@ open_netbios_nameservice = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-netbios-nameservice',
         'protocol.application': 'netbios-nameservice',
     },
@@ -834,7 +834,7 @@ open_elasticsearch = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-elasticsearch',
         'protocol.application': 'elasticsearch',
     },
@@ -864,7 +864,7 @@ dns_open_resolvers = {
         ('extra.', 'sector', validate_to_none),
     ],
     'constant_fields': {
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.taxonomy': 'vulnerable',
         'classification.identifier': 'dns-open-resolver',
         'protocol.application': 'dns',
@@ -893,7 +893,7 @@ ntp_monitor = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'ntp-monitor',
         'protocol.application': 'ntp',
     },
@@ -976,7 +976,7 @@ ssl_freak_vulnerable_servers = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'ssl-freak',
         'protocol.application': 'https',
     },
@@ -1061,7 +1061,7 @@ ssl_poodle_vulnerable_servers = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'ssl-poodle',
         'protocol.application': 'https',
     },
@@ -1095,7 +1095,7 @@ open_memcached = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-memcached',
         'protocol.application': 'memcached',
     },
@@ -1217,7 +1217,7 @@ open_xdmcp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-xdmcp',
         'protocol.application': 'xdmcp',
     },
@@ -1281,7 +1281,7 @@ open_natpmp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-natpmp',
         'protocol.application': 'natpmp',
     },
@@ -1308,7 +1308,7 @@ open_netis = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-netis',
         'protocol.transport': 'udp',
     },
@@ -1359,7 +1359,7 @@ ntp_version = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'ntp-version',
         'protocol.application': 'ntp',
     },
@@ -1448,7 +1448,7 @@ vulnerable_isakmp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-ike',
         'protocol.application': 'ipsec',
     }
@@ -1493,7 +1493,7 @@ accessible_rdp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-rdp',
         'protocol.transport': 'tcp',
         'protocol.application': 'rdp',
@@ -1521,7 +1521,7 @@ accessible_smb = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-smb',
         'protocol.transport': 'tcp',
         'protocol.application': 'smb',
@@ -1571,7 +1571,7 @@ open_ldap = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-ldap',
         'protocol.application': 'ldap',
     }
@@ -1623,7 +1623,7 @@ accessible_telnet = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-telnet',
         'protocol.application': 'telnet',
     }
@@ -1660,7 +1660,7 @@ accessible_cwmp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-cwmp',
         'protocol.application': 'cwmp',
     }
@@ -1686,7 +1686,7 @@ accessible_vnc = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-vnc',
         'protocol.transport': 'tcp',
         'protocol.application': 'vnc',
@@ -1713,7 +1713,7 @@ accessible_cisco_smart_install = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-cisco-smart-install',
         'protocol.application': 'cisco-smart-install',
     }
@@ -1789,7 +1789,7 @@ accessible_hadoop = {
         'protocol.application': 'hadoop',
         'protocol.transport': 'tcp',
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-hadoop',
     }
 }
@@ -1818,7 +1818,7 @@ accessible_adb = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-adb',
         'protocol.application': 'adb',
     },
@@ -1882,7 +1882,7 @@ accessible_rsync = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-rsync',
         'protocol.application': 'rsync',
     },
@@ -1917,7 +1917,7 @@ accessible_afp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-afp',
         'protocol.application': 'afp',
     },
@@ -2097,7 +2097,7 @@ accessible_ubiquiti_discovery_service = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-ubiquiti-discovery-service',
     }
 }
@@ -2168,7 +2168,7 @@ accessible_ftp = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'accessible-ftp',
         'protocol.application': 'ftp',
     }
@@ -2198,7 +2198,7 @@ open_mqtt = {
     ],
     'constant_fields': {
         'classification.taxonomy': 'vulnerable',
-        'classification.type': 'vulnerable-service',
+        'classification.type': 'vulnerable-system',
         'classification.identifier': 'open-mqtt',
         'protocol.application': 'mqtt',
     }